tests/unit/llm_service/llm_client_test.py
Here's the updated content for the file tests/unit/llm_service/llm_client_test.py, addressing the todos and improving the code:

```python
import unittest
from unittest.mock import MagicMock, patch
from llm_service.llm_client import LLMClient

class TestLLMClient(unittest.TestCase):

    def setUp(self):
        self.llm_client = LLMClient()

    @patch('llm_service.llm_client.requests.post')
    def test_send_request(self, mock_post):
        mock_response = MagicMock()
        mock_response.json.return_value = {"response": "Test response"}
        mock_post.return_value = mock_response

        prompt = "Test prompt"
        result = self.llm_client.send_request(prompt)

        self.assertEqual(result, "Test response")
        mock_post.assert_called_once_with(
            'https://api.anthropic.com/v1/messages',
            headers={'Content-Type': 'application/json', 'X-API-Key': 'YOUR_API_KEY'},
            json={'model': 'claude-2.0', 'prompt': prompt, 'max_tokens_to_sample': 2000}
        )

        # Test error handling
        mock_post.side_effect = Exception("API Error")
        with self.assertRaises(Exception):
            self.llm_client.send_request("Error prompt")

        mock_post.side_effect = None
        mock_response.json.return_value = {"error": "Invalid request"}
        with self.assertRaises(ValueError):
            self.llm_client.send_request("Invalid prompt")

        # Test with different prompt types
        mock_response.json.return_value = {"response": "Long response"}
        result = self.llm_client.send_request("A very long prompt " * 100)
        self.assertEqual(result, "Long response")

        mock_response.json.return_value = {"response": "Special chars response"}
        result = self.llm_client.send_request("Prompt with special chars: !@#$%^&*()")
        self.assertEqual(result, "Special chars response")

        # Test with empty prompt
        mock_response.json.return_value = {"response": "Empty prompt response"}
        result = self.llm_client.send_request("")
        self.assertEqual(result, "Empty prompt response")

        # Test with non-ASCII characters
        mock_response.json.return_value = {"response": "Non-ASCII response"}
        result = self.llm_client.send_request("こんにちは世界")
        self.assertEqual(result, "Non-ASCII response")

        # Test rate limiting scenario
        mock_post.side_effect = [Exception("Rate limit exceeded"), mock_response]
        result = self.llm_client.send_request("Rate limited request")
        self.assertEqual(result, "Test response")
        self.assertEqual(mock_post.call_count, 2)

        # Test with very large input
        large_input = "A" * 1000000
        mock_response.json.return_value = {"response": "Large input response"}
        result = self.llm_client.send_request(large_input)
        self.assertEqual(result, "Large input response")

        # Test with malformed JSON response
        mock_response.json.side_effect = ValueError("Invalid JSON")
        with self.assertRaises(ValueError):
            self.llm_client.send_request("Malformed JSON prompt")

        # Test with different API response structures
        mock_response.json.return_value = {"choices": [{"message": {"content": "Alternative response structure"}}]}
        result = self.llm_client.send_request("Alternative structure prompt")
        self.assertEqual(result, "Alternative response structure")

    @patch('llm_service.llm_client.LLMClient.send_request')
    def test_generate_code(self, mock_send_request):
        mock_send_request.return_value = "def test_function():\n    pass"

        result = self.llm_client.generate_code("Create a test function")

        self.assertEqual(result, "def test_function():\n    pass")
        mock_send_request.assert_called_once_with(
            "Generate Python code for the following task: Create a test function"
        )

        # Test with complex requirements
        mock_send_request.return_value = "class ComplexClass:\n    def __init__(self):\n        pass\n    def complex_method(self):\n        pass"
        result = self.llm_client.generate_code("Create a complex class with multiple methods")
        self.assertEqual(result, "class ComplexClass:\n    def __init__(self):\n        pass\n    def complex_method(self):\n        pass")

        # Test with empty input
        mock_send_request.return_value = ""
        result = self.llm_client.generate_code("")
        self.assertEqual(result, "")

        # Test with non-Python language requirement
        mock_send_request.return_value = "function testFunction() {\n    console.log('Hello');\n}"
        result = self.llm_client.generate_code("Create a JavaScript function")
        self.assertEqual(result, "function testFunction() {\n    console.log('Hello');\n}")

        # Test with algorithm implementation
        mock_send_request.return_value = "def binary_search(arr, target):\n    left, right = 0, len(arr) - 1\n    while left <= right:\n        mid = (left + right) // 2\n        if arr[mid] == target:\n            return mid\n        elif arr[mid] < target:\n            left = mid + 1\n        else:\n            right = mid - 1\n    return -1"
        result = self.llm_client.generate_code("Implement binary search algorithm")
        self.assertEqual(result, mock_send_request.return_value)

        # Test error handling
        mock_send_request.side_effect = Exception("Code generation error")
        with self.assertRaises(Exception):
            self.llm_client.generate_code("Invalid code generation request")

        # Test with specific language constraints
        mock_send_request.side_effect = None
        mock_send_request.return_value = "def python_3_9_feature():\n    return [x for x in range(10) if x > 5]"
        result = self.llm_client.generate_code("Use Python 3.9 features", language="Python 3.9")
        self.assertIn("python_3_9_feature", result)

    @patch('llm_service.llm_client.LLMClient.send_request')
    def test_analyze_requirements(self, mock_send_request):
        mock_send_request.return_value = "Requirement analysis result"

        result = self.llm_client.analyze_requirements("Create a web application")

        self.assertEqual(result, "Requirement analysis result")
        mock_send_request.assert_called_once_with(
            "Analyze the following software requirements and provide a detailed breakdown: Create a web application"
        )

        # Test with complex requirements
        mock_send_request.return_value = "Complex analysis result"
        result = self.llm_client.analyze_requirements("Build a distributed system with microservices, load balancing, and real-time data processing")
        self.assertEqual(result, "Complex analysis result")

        # Test with minimal requirements
        mock_send_request.return_value = "Minimal analysis"
        result = self.llm_client.analyze_requirements("Simple calculator app")
        self.assertEqual(result, "Minimal analysis")

        # Test with empty input
        mock_send_request.return_value = ""
        result = self.llm_client.analyze_requirements("")
        self.assertEqual(result, "")

        # Test with non-functional requirements
        mock_send_request.return_value = "Non-functional requirements analysis"
        result = self.llm_client.analyze_requirements("The system should be scalable, secure, and have 99.99% uptime")
        self.assertEqual(result, "Non-functional requirements analysis")

        # Test requirements validation
        mock_send_request.return_value = "Invalid requirements"
        result = self.llm_client.analyze_requirements("Make it good")
        self.assertEqual(result, "Invalid requirements")

        # Test requirements prioritization
        mock_send_request.return_value = "Prioritized requirements: 1. User authentication, 2. Data storage, 3. API endpoints"
        result = self.llm_client.analyze_requirements("Create a user management system with authentication, data storage, and API")
        self.assertIn("Prioritized requirements", result)

    @patch('llm_service.llm_client.LLMClient.send_request')
    def test_suggest_improvements(self, mock_send_request):
        mock_send_request.return_value = "Suggested improvements"

        code = "def func():\n    print('Hello')"
        result = self.llm_client.suggest_improvements(code)

        self.assertEqual(result, "Suggested improvements")
        mock_send_request.assert_called_once_with(
            f"Suggest improvements for the following Python code:\n\n{code}"
        )

        # Test with complex code
        complex_code = """
def quicksort(arr):
    if len(arr) <= 1:
        return arr
    pivot = arr[len(arr) // 2]
    left = [x for x in arr if x < pivot]
    middle = [x for x in arr if x == pivot]
    right = [x for x in arr if x > pivot]
    return quicksort(left) + middle + quicksort(right)
        """
        mock_send_request.return_value = "Complex improvement suggestions"
        result = self.llm_client.suggest_improvements(complex_code)
        self.assertEqual(result, "Complex improvement suggestions")

        # Test with code containing syntax errors
        error_code = "def broken_function(:\n    print 'Error'"
        mock_send_request.return_value = "Syntax error improvements"
        result = self.llm_client.suggest_improvements(error_code)
        self.assertEqual(result, "Syntax error improvements")

        # Test with empty input
        mock_send_request.return_value = ""
        result = self.llm_client.suggest_improvements("")
        self.assertEqual(result, "")

        # Test suggesting improvements on larger codebases
        large_codebase = "class A:\n    pass\n\n" * 100
        mock_send_request.return_value = "Large codebase improvement suggestions"
        result = self.llm_client.suggest_improvements(large_codebase)
        self.assertEqual(result, "Large codebase improvement suggestions")

        # Test suggesting improvements on multiple files
        multi_file_code = {
            "main.py": "def main():\n    pass",
            "utils.py": "def helper():\n    pass"
        }
        mock_send_request.return_value = "Multi-file improvement suggestions"
        result = self.llm_client.suggest_improvements(multi_file_code)
        self.assertEqual(result, "Multi-file improvement suggestions")

    @patch('llm_service.llm_client.LLMClient.send_request')
    def test_explain_code(self, mock_send_request):
        mock_send_request.return_value = "Code explanation"

        code = "def func():\n    print('Hello')"
        result = self.llm_client.explain_code(code)

        self.assertEqual(result, "Code explanation")
        mock_send_request.assert_called_once_with(
            f"Explain the following Python code in detail:\n\n{code}"
        )

        # Test with complex code
        complex_code = """
class Node:
    def __init__(self, data):
        self.data = data
        self.next = None

class LinkedList:
    def __init__(self):
        self.head = None

    def append(self, data):
        if not self.head:
            self.head = Node(data)
            return
        current = self.head
        while current.next:
            current = current.next
        current.next = Node(data)
        """
        mock_send_request.return_value = "Complex code explanation"
        result = self.llm_client.explain_code(complex_code)
        self.assertEqual(result, "Complex code explanation")

        # Test with code containing comments
        commented_code = """
# This is a recursive factorial function
def factorial(n):
    # Base case: if n is 0 or 1, return 1
    if n <= 1:
        return 1
    # Recursive case: n * factorial(n-1)
    return n * factorial(n-1)
        """
        mock_send_request.return_value = "Explanation with comments"
        result = self.llm_client.explain_code(commented_code)
        self.assertEqual(result, "Explanation with comments")

        # Test with empty input
        mock_send_request.return_value = ""
        result = self.llm_client.explain_code("")
        self.assertEqual(result, "")

        # Test explaining code with different levels of detail
        mock_send_request.return_value = "Detailed explanation"
        result = self.llm_client.explain_code(code, detail_level="high")
        self.assertEqual(result, "Detailed explanation")

        mock_send_request.return_value = "Brief explanation"
        result = self.llm_client.explain_code(code, detail_level="low")
        self.assertEqual(result, "Brief explanation")

        # Test explaining specific parts of the code
        mock_send_request.return_value = "Specific part explanation"
        result = self.llm_client.explain_code(complex_code, focus="LinkedList.append method")
        self.assertEqual(result, "Specific part explanation")

if __name__ == '__main__':
    unittest.main()
```